+++ apps/angular-feature-sliced/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["../../.eslintrc.json"],
+  "extends": ["../../.eslintrc.json", "@feature-sliced"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
