+++ apps/angular-feature-sliced/src/app/app.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { AppRoutingModule } from 'pages';
 import { AppComponent } from 'app/app.component';
 import { AppStoreModule } from 'app/core';
-import { AppRoutingModule } from 'pages';
 
 @NgModule({
   declarations: [AppComponent],
