+++ apps/angular-feature-sliced/src/app/core/store/store.ts
@@ -1,8 +1,8 @@
 import { InjectionToken, Type } from '@angular/core';
 import * as fromRouter from '@ngrx/router-store';
 import { Action, ActionReducerMap, MetaReducer } from '@ngrx/store';
+import { environment } from 'environments';
 import { taskModel } from 'entities/task';
-import { environment } from 'environments/environment';
 
 /**
  * Every reducer module's default export is the reducer function it¬self. In
