+++ apps/angular-feature-sliced/src/environments/index.ts
@@ -1,0 +1,1 @@
+export { environment } from './environment';
