+++ apps/angular-feature-sliced/src/features/toggle-task/components/toggle-task/toggle-task.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-import type { Task } from 'shared/api';
 import { TaskFacade } from 'entities/task';
+import type { Task } from 'shared/api';
 
 @Component({
   selector: 'fs-toggle-task',
