+++ apps/angular-feature-sliced/src/pages/task-details/task-details.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
-import { TaskModule } from 'entities/task';
 import { SharedModule, UiKitModule } from 'shared';
 import { ToggleTaskModule } from 'features/toggle-task';
+import { TaskModule } from 'entities/task';
 
 import { TaskDetailsRoutingModule } from './task-details-routing.module';
 import { TaskDetailsPage } from './task-details.page';
